requirements.txt
openai
python-dotenv
requests
google-search-results